Change: objects -> items
More concise, same meaning--what's not to love?

In the documentation, the term "object" is still used in places to
subtly distinguish between arbitrary objects (which might be added to
a taxy's items) and ones that have been added to a taxy's
items (thereby becoming one of that taxy's items).  Yay, jargon.