Fix: (taxy-magit-section-define-column-definer) Don't pass ellipsis
Don't pass string as ELLIPSIS argument to TRUNCATE-STRING-TO-WIDTH;
instead pass t, which defaults to value of variable
TRUNCATE-STRING-ELLIPSIS.

The width of the string passed before, the U+2026 HORIZONTAL ELLIPSIS
character, varies by font, and fonts which display it with a width
different than that of a single space cause misalignment of columns.

Now users may specify the ellipsis string according to their needs.