Fix: (-format-items) Calculating image widths if on a text frame
Try to find a graphical frame when calculating image widths for column
widths (or signal an error if none are available, rather than leaving
IMAGE-SIZE to signal an error).

For example, if both graphical and text frames are available, and a
column's values contain an image, and the buffer is being redisplayed
on a text frame, try to use a graphical frame for calculating the
image width, rather than just signaling an error.

If only text frames are available, images should not be present in
strings to begin with, so we don't try to handle that case.