WIP
This seems to improve performance of redrawing the room list
slightly (when used with the matching branch of taxy-magit-section),
but there's still much work done that isn't cached.  Really, the whole
taxy-magit-section-format-items function should be rewritten so that
the labeled functions aren't closures and don't use side effects (even
though it's nifty, it makes caching harder), and can cache their
values.