Fix: (taxy-magit-section-format-items) Only use visible frames
When server-mode is enabled, an invisible frame is present, which
still returns X from FRAMEP, but which should not be used for
calculating image sizes.

This should fix an "Invalid image size" error that was happening (when
such an invisible frame happened to be sorted first in FRAME-LIST).

See <https://github.com/alphapapa/ement.el/issues/298#issuecomment-2369652629>.