Fix: (taxy-magit-section-insert) Accept taxy struct
The documented examples suggest that taxy-magit-section-pp can accept
a taxy struct.  However, taxy-magit-section-insert's labeled function
insert-taxy attempts to access the level-indent slot of the taxy which
was passed in.

Now, taxy-magit-section-insert's labeled function insert-taxy converts
taxy structs to taxy-magit-section structs.  Examples in README.org
which call taxy-magit-section-pp now work as-is.